+++ python/cugraph/__init__.py
@@ -32,6 +32,7 @@ from cugraph.structure import (
     hypergraph,
     symmetrize,
     symmetrize_df,
+    symmetrize_ddf,
 )
 
 from cugraph.centrality import (